Allow optimization level on a per-query level
