+++ bench/benches/bsbm_explore.rs
@@ -5,11 +5,11 @@
 //!
 //! The tests assume the presence of the benchmark data.
 
-use codspeed_criterion_compat::{criterion_group, criterion_main, Criterion};
+use codspeed_criterion_compat::{Criterion, criterion_group, criterion_main};
 use futures::StreamExt;
 use rdf_fusion::io::RdfFormat;
 use rdf_fusion::store::Store;
-use rdf_fusion::{QueryOptions, QueryResults};
+use rdf_fusion::{OptimizationLevel, QueryOptions, QueryResults};
 use std::fs;
 use std::path::PathBuf;
 use tokio::runtime::{Builder, Runtime};
