+++ lib/execution/src/sparql/mod.rs
@@ -4,6 +4,7 @@ mod algebra;
 pub mod error;
 mod eval;
 mod explanation;
+mod optimizer;
 mod rewriting;
 
 pub use crate::results::{
@@ -16,15 +17,19 @@ pub use rdf_fusion_model::{Variable, VariableNameParseError};
 pub use spargebra::SparqlSyntaxError;
 
 /// Defines how many optimizations the query optimizer should apply.
+///
+/// Currently, the default value is [OptimizationLevel::Full], as we are still searching for a
+/// subset that performs well on many queries. Once this subset has been identified, the default
+/// value will be [OptimizationLevel::Default].
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
 pub enum OptimizationLevel {
     /// No optimizations, except rewrites that are necessary for a working query.
     None,
     /// A balanced default optimization level. Suitable for simple queries or those handling modest
     /// data volumes.
-    #[default]
     Default,
     /// Runs all optimizations. Ideal for complex queries or those processing large datasets.
+    #[default]
     Full,
 }
